--- a/ggml/src/ggml-vulkan/vulkan-shaders/types.comp
+++ b/ggml/src/ggml-vulkan/vulkan-shaders/types.comp
@@ -468,9 +468,11 @@ void init_iq_shmem(uvec3 wgsize)
     // copy the table into shared memory and sync
     [[unroll]] for (uint i = 0; i < iq1s_grid_const.length(); i += wgsize.x) {
         uint idx = i + gl_LocalInvocationIndex.x;
-        u16vec2 g = unpack16(iq1s_grid_const[idx]);
-        iq1s_grid[2*idx+0] = g.x;
-        iq1s_grid[2*idx+1] = g.y;
+        if (iq1s_grid_const.length() % wgsize.x == 0 || idx < iq1s_grid_const.length()) {
+            u16vec2 g = unpack16(iq1s_grid_const[idx]);
+            iq1s_grid[2*idx+0] = g.x;
+            iq1s_grid[2*idx+1] = g.y;
+        }
     }
     barrier();
 }
@@ -567,7 +569,9 @@ void init_iq_shmem(uvec3 wgsize)
 {
     // copy the table into shared memory and sync
     [[unroll]] for (uint i = 0; i < iq2xxs_grid.length(); i += wgsize.x) {
-        iq2xxs_grid[i + gl_LocalInvocationIndex.x] = iq2xxs_grid_const[i + gl_LocalInvocationIndex.x];
+        if (iq2xxs_grid_const.length() % wgsize.x == 0 || i + gl_LocalInvocationIndex.x < iq2xxs_grid_const.length()) {
+            iq2xxs_grid[i + gl_LocalInvocationIndex.x] = iq2xxs_grid_const[i + gl_LocalInvocationIndex.x];
+        }
     }
     barrier();
 }
@@ -735,7 +739,9 @@ void init_iq_shmem(uvec3 wgsize)
 {
     // copy the table into shared memory and sync
     [[unroll]] for (uint i = 0; i < iq2xs_grid.length(); i += wgsize.x) {
-        iq2xs_grid[i + gl_LocalInvocationIndex.x] = iq2xs_grid_const[i + gl_LocalInvocationIndex.x];
+        if (iq2xs_grid.length() % wgsize.x == 0 || i + gl_LocalInvocationIndex.x < iq2xs_grid_const.length()) {
+            iq2xs_grid[i + gl_LocalInvocationIndex.x] = iq2xs_grid_const[i + gl_LocalInvocationIndex.x];
+        }
     }
     barrier();
 }
@@ -1033,7 +1039,9 @@ void init_iq_shmem(uvec3 wgsize)
 {
     // copy the table into shared memory and sync
     [[unroll]] for (uint i = 0; i < iq2s_grid.length(); i += wgsize.x) {
-        iq2s_grid[i + gl_LocalInvocationIndex.x] = iq2s_grid_const[i + gl_LocalInvocationIndex.x];
+        if (iq2s_grid.length() % wgsize.x == 0 || i + gl_LocalInvocationIndex.x < iq2s_grid_const.length()) {
+            iq2s_grid[i + gl_LocalInvocationIndex.x] = iq2s_grid_const[i + gl_LocalInvocationIndex.x];
+        }
     }
     barrier();
 }
@@ -1103,7 +1111,9 @@ void init_iq_shmem(uvec3 wgsize)
 {
     // copy the table into shared memory and sync
     [[unroll]] for (uint i = 0; i < iq3xxs_grid.length(); i += wgsize.x) {
-        iq3xxs_grid[i + gl_LocalInvocationIndex.x] = iq3xxs_grid_const[i + gl_LocalInvocationIndex.x];
+        if (iq3xxs_grid.length() % wgsize.x == 0 || i + gl_LocalInvocationIndex.x < iq3xxs_grid.length()) {
+            iq3xxs_grid[i + gl_LocalInvocationIndex.x] = iq3xxs_grid_const[i + gl_LocalInvocationIndex.x];
+        }
     }
     barrier();
 }
@@ -1211,7 +1221,9 @@ void init_iq_shmem(uvec3 wgsize)
 {
     // copy the table into shared memory and sync
     [[unroll]] for (uint i = 0; i < iq3s_grid.length(); i += wgsize.x) {
-        iq3s_grid[i + gl_LocalInvocationIndex.x] = iq3s_grid_const[i + gl_LocalInvocationIndex.x];
+        if (iq3s_grid.length() % wgsize.x == 0 || i + gl_LocalInvocationIndex.x < iq3s_grid.length()) {
+            iq3s_grid[i + gl_LocalInvocationIndex.x] = iq3s_grid_const[i + gl_LocalInvocationIndex.x];
+        }
     }
     barrier();
 }