@jeffbolznv 2025-02-02
> llvmpipe seems to have issues with the `init_iq_shmem` of `iq2_xxs`
> 
> adding a bounds check makes it happy

I didn't realize we were using such large workgroup sizes with these init functions for getrows. Maybe the branch condition should do something like `((length % wgsize.x) != 0) &&` so it's optimized away in the mul mat shaders.