@remyoudompheng 2025-02-15
It should be OK now, llvmpipe seems happy.