@jeffbolznv 2025-02-16
I've verified all tests are passing with the latest commit on RTX 4070.