@alexjp 2025-02-18
I am sorry if this is offtopic, but seeing that it was due to this pull request and using this branch, but may I ask if this is normal:

I did a simple benchmark, using a 3B llm model, and these were the results:

```
IQ4_NL = 60.4 t/s
IQ4_XS = 36.4 t/s
IQ3_M = 57.9 t/s
IQ3_S = 57.4 t/s
IQ3_XS = 60.4 t/s
IQ2_M = 60.0 t/s
```

is it expected for IQ4_XS to be so slow?